requirements.txt
albucore==0.0.12
albumentations==1.4.10
annotated-types==0.7.0
bleach==6.1.0
bokeh==3.4.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorcet==3.1.0
contourpy==1.2.1
cycler==0.12.1
docker-pycreds==0.4.0
efficientnet-pytorch==0.7.1
filelock==3.15.4
fonttools==4.53.0
fsspec==2024.6.0
ftfy==6.2.0
gitdb==4.0.11
gitpython==3.1.43
holoviews==1.19.0
huggingface-hub==0.23.4
hvplot==0.10.0
idna==3.7
imageio==2.34.2
jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lazy-loader==0.4
linkify-it-py==2.0.3
markdown==3.6
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.9.0
mdit-py-plugins==0.4.1
mdurl==0.1.2
mpmath==1.3.0
munch==4.0.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.2
panel==1.4.4
param==2.1.1
pie-torch==0.0.1
pillow==10.3.0
platformdirs==4.2.2
polars==0.20.31
pretrainedmodels==0.7.4
protobuf==5.27.2
psutil==6.0.0
pydantic==2.7.4
pydantic-core==2.18.4
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyviz-comms==3.0.2
pyyaml==6.0.1
regex==2024.5.15
requests==2.32.3
rootutils==1.0.7
safetensors==0.4.3
scikit-image==0.24.0
scikit-learn==1.5.0
scipy==1.14.0
sentry-sdk==2.7.0
setproctitle==1.3.3
setuptools==69.5.1
six==1.16.0
smmap==5.0.1
sympy==1.12.1
threadpoolctl==3.5.0
tifffile==2024.6.18
timm==1.0.7
tokenizers==0.19.1
tomli==2.0.1
torch==2.3.1
torchshow==0.5.1
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.4
transformers==4.41.2
triton==2.3.1
typing-extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
urllib3==2.2.2
wandb==0.17.3
wcwidth==0.2.13
webencodings==0.5.1
xyzservices==2024.6.0
